util: move unique integer generation logic to a util package
This is a mechanical change to move the logic for generating unique
64-bit IDs to `pkg/util/unique`. This allows users to avoid depending
on `pkg/sql/sem/builtins`.

Epic: None

Release note: None
