+++ pkg/BUILD.bazel
@@ -778,6 +778,7 @@ ALL_TESTS = [
     "//pkg/util/tsearch:tsearch_test",
     "//pkg/util/uint128:uint128_test",
     "//pkg/util/ulid:ulid_test",
+    "//pkg/util/unique:unique_test",
     "//pkg/util/uuid:uuid_test",
     "//pkg/util/vector:vector_test",
     "//pkg/util/version:version_test",
@@ -2670,6 +2671,8 @@ GO_TARGETS = [
     "//pkg/util/ulid:ulid",
     "//pkg/util/ulid:ulid_test",
     "//pkg/util/unaccent:unaccent",
+    "//pkg/util/unique:unique",
+    "//pkg/util/unique:unique_test",
     "//pkg/util/uuid:uuid",
     "//pkg/util/uuid:uuid_test",
     "//pkg/util/vector:vector",
