+++ pkg/backup/backupsink/BUILD.bazel
@@ -19,11 +19,11 @@ go_library(
         "//pkg/roachpb",
         "//pkg/settings",
         "//pkg/sql/execinfrapb",
-        "//pkg/sql/sem/builtins",
         "//pkg/storage",
         "//pkg/util/admission",
         "//pkg/util/hlc",
         "//pkg/util/log",
+        "//pkg/util/unique",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_gogo_protobuf//types",
         "@com_github_kr_pretty//:pretty",
