+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -118,7 +118,6 @@ go_library(
         "//pkg/sql/rowenc",
         "//pkg/sql/rowexec",
         "//pkg/sql/sem/asof",
-        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/catid",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
@@ -164,6 +163,7 @@ go_library(
         "//pkg/util/timeofday",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "@com_github_apache_pulsar_client_go//pulsar",
         "@com_github_cockroachdb_apd_v3//:apd",
