+++ pkg/crosscluster/producer/BUILD.bazel
@@ -52,7 +52,6 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/privilege",
-        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
         "//pkg/sql/syntheticprivilege",
@@ -68,6 +67,7 @@ go_library(
         "//pkg/util/span",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_logtags//:logtags",
