+++ pkg/jobs/BUILD.bazel
@@ -54,7 +54,6 @@ go_library(
         "//pkg/sql/isql",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/protoreflect",
-        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/catconstants",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
@@ -77,6 +76,7 @@ go_library(
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
