+++ pkg/sql/sem/builtins/BUILD.bazel
@@ -32,7 +32,6 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sem/builtins",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/base",
         "//pkg/build",
         "//pkg/clusterversion",
         "//pkg/config/zonepb",
@@ -136,6 +135,7 @@ go_library(
         "//pkg/util/tsearch",
         "//pkg/util/ulid",
         "//pkg/util/unaccent",
+        "//pkg/util/unique",
         "//pkg/util/uuid",
         "//pkg/util/vector",
         "@com_github_cockroachdb_apd_v3//:apd",
