+++ pkg/sql/sem/builtins/builtinconstants/constants.go
@@ -71,29 +71,3 @@ const (
 	// creates a job that logs SQL schema telemetry.
 	CreateSchemaTelemetryJobBuiltinName = "crdb_internal.create_sql_schema_telemetry_job"
 )
-
-// A unique int generated by GenerateUniqueInt is a 64-bit integer with
-// the following format:
-//
-//	[1 leading zero bit][48 bits for timestamp][15 bits for nodeID]
-const (
-	// UniqueIntLeadingZeroBits is the number of leading zero bits in a unique
-	// int generated by GenerateUniqueInt.
-	UniqueIntLeadingZeroBits = 1
-
-	// UniqueIntTimestampBits is the number of bits in the timestamp segment
-	// in a unique int generated by GenerateUniqueInt.
-	UniqueIntTimestampBits = 48
-
-	// UniqueIntNodeIDBits is the number of bits in the node ID segment
-	// in a unique int generated by GenerateUniqueInt.
-	UniqueIntNodeIDBits = 15
-
-	// UniqueIntNodeIDMask is a bitmask for the node ID in a unique int
-	// generated by GenerateUniqueInt.
-	UniqueIntNodeIDMask = 1<<UniqueIntNodeIDBits - 1
-
-	// UniqueIntTimestampMask is a bitmask for the timestamp in a unique int
-	// generated by GenerateUniqueInt.
-	UniqueIntTimestampMask = (1<<UniqueIntTimestampBits - 1) << UniqueIntNodeIDBits
-)
