+++ pkg/sql/sem/builtins/math_builtins.go
@@ -8,7 +8,6 @@ package builtins
 import (
 	"context"
 	"math"
-	"time"
 
 	"github.com/cockroachdb/apd/v3"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
@@ -17,7 +16,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/volatility"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
-	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/errors"
 )
 
@@ -771,13 +769,6 @@ func roundDecimal(x *apd.Decimal, scale int32) (tree.Datum, error) {
 	}
 	return dd, err
 }
-
-var uniqueIntState struct {
-	syncutil.Mutex
-	timestamp uint64
-}
-
-var uniqueIntEpoch = time.Date(2015, time.January, 1, 0, 0, 0, 0, time.UTC).UnixNano()
 
 // widthBucket returns the bucket number to which operand would be assigned in a histogram having count
 // equal-width buckets spanning the range b1 to b2
