+++ pkg/sql/vecindex/vecstore/BUILD.bazel
@@ -56,7 +56,6 @@ go_library(
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
         "//pkg/sql/rowinfra",
-        "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/tree",
         "//pkg/sql/vecindex/internal",
         "//pkg/sql/vecindex/quantize",
@@ -66,6 +65,7 @@ go_library(
         "//pkg/util/encoding",
         "//pkg/util/protoutil",
         "//pkg/util/syncutil",
+        "//pkg/util/unique",
         "//pkg/util/vector",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_gogo_protobuf//gogoproto",
